requirements.txt
Django                 2.1.5
django-rest-auth       0.9.3
djangorestframework    3.8.2
oauthlib               3.0.1
pip                    18.0
requests               2.21.0
requests-oauthlib      1.2.0
setuptools             40.4.3
social-auth-app-django 3.1.0
social-auth-core       3.0.0
urllib3                1.24.1
virtualenv             16.0.0
virtualenvwrapper-win  1.2.5